requirements.txt
Flask==2.3.3
Flask-Migrate==2.5.2
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.2
flask-swagger-ui==4.11.1
Pillow==10.0.1
psycopg2-binary==2.9.7
PyJWT==2.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.21
virtualenv==20.24.5
Werkzeug==2.3.7
flask-script
